Removed nx from most traversal/test_sssp
